Rimosso cast problematico

+++ main/utils/utils.h
@@ -5,7 +5,7 @@
 #include "freertos/task.h"
 #include "lvgl.h"
 
-#define get_millis() ((uint32_t)xTaskGetTickCount() * portTICK_PERIOD_MS)
+#define get_millis() (xTaskGetTickCount() * portTICK_PERIOD_MS)
 void free_lv_user_data(lv_obj_t *obj);
 
 #endif
